refactor: ui and keycode
