+++ src/main.rs
@@ -1,3 +1,4 @@
+mod ui;
 use crossterm::{
     event::{self, Event, KeyCode},
     execute,
@@ -7,14 +8,8 @@ use octocrab::models::pulls::PullRequest as OctocrabPullRequest;
 use octocrab::Error as OctocrabError;
 use octocrab::Octocrab;
 
-use ratatui::{
-    backend::CrosstermBackend,
-    layout::{Constraint, Direction, Layout, Rect},
-    style::{Color, Style},
-    text::{Line, Span, Text},
-    widgets::{Block, BorderType, Borders, Clear, Paragraph, Wrap},
-    Frame, Terminal,
-};
+use crate::ui::layout::ui;
+use ratatui::{backend::CrosstermBackend, Terminal};
 use std::io;
 use thiserror::Error;
 
@@ -79,9 +74,9 @@ impl App {
         }
     }
 
-    fn enter_edit_mode(&mut self) {
+    fn enter_edit_mode(&mut self, index: usize) {
         self.input_mode = InputMode::Editing;
-        self.current_field = 0;
+        self.current_field = index;
     }
 
     fn preview_pull_request(&mut self) {
@@ -99,6 +94,8 @@ impl App {
         self.input_mode = InputMode::Normal;
         self.current_field = 0;
         self.show_popup = false;
+        self.error_message = None;
+        self.success_message = None;
     }
 
     async fn create_github_pull_request(&self) -> Result<OctocrabPullRequest, PullRequestError> {
@@ -155,11 +152,6 @@ impl App {
     fn set_success(&mut self, success: String) {
         self.success_message = Some(success);
     }
-
-    fn clear_result(&mut self) {
-        self.error_message = None;
-        self.success_message = None;
-    }
 }
 
 fn main() -> Result<(), io::Error> {
@@ -179,10 +171,11 @@ fn main() -> Result<(), io::Error> {
                 InputMode::Normal => match key.code {
                     KeyCode::Char('q') => break,
                     KeyCode::Char('e') => {
-                        app.enter_edit_mode();
+                        app.enter_edit_mode(app.current_field);
                     }
-                    KeyCode::Char('c') => {
-                        app.preview_pull_request();
+                    KeyCode::Char('n') => {
+                        app.reset();
+                        app.enter_edit_mode(0);
                     }
                     KeyCode::Down => {
                         app.current_field = (app.current_field + 1) % 4;
@@ -209,19 +202,21 @@ fn main() -> Result<(), io::Error> {
                         let current_field = app.get_current_field_mut();
                         if current_field_index == 1 {
                             current_field.push('\n');
-                        } else if current_field_index == 3 {
-                            app.preview_pull_request();
                         } else {
-                            app.current_field = (app.current_field + 1) % 4;
+                            app.preview_pull_request();
                         }
                     }
                     KeyCode::Tab => {
                         app.current_field = (app.current_field + 1) % 4;
                     }
+                    KeyCode::BackTab => {
+                        app.current_field = (app.current_field + 3) % 4;
+                    }
                     _ => {}
                 },
                 InputMode::Creating => match key.code {
                     KeyCode::Enter => {
+                        app.input_mode = InputMode::Normal;
                         app.show_popup = false;
                         let result = tokio::runtime::Runtime::new()
                             .unwrap()
@@ -248,9 +243,9 @@ fn main() -> Result<(), io::Error> {
                         app.input_mode = InputMode::Editing;
                         app.show_popup = false;
                     }
-                    KeyCode::Char('n') => {
-                        app.reset();
-                        app.clear_result();
+                    KeyCode::Char('q') => {
+                        app.input_mode = InputMode::Normal;
+                        app.show_popup = false;
                     }
                     _ => {}
                 },
@@ -263,200 +258,4 @@ fn main() -> Result<(), io::Error> {
     terminal.show_cursor()?;
 
     Ok(())
-}
-
-fn ui(f: &mut Frame, app: &App) {
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(1)
-        .constraints(
-            [
-                Constraint::Percentage(5),
-                Constraint::Percentage(40),
-                Constraint::Percentage(50),
-                Constraint::Percentage(5),
-            ]
-            .as_ref(),
-        )
-        .split(f.area());
-
-    let block = Block::default()
-        .title("PRT: Pull Request TUI")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded);
-    f.render_widget(block, f.area());
-
-    let repo_block = Block::default()
-        .title("Context")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded);
-
-    let repo_area = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(
-            [
-                Constraint::Percentage(100), // Ajusta según el tamaño que prefieras
-            ]
-            .as_ref(),
-        )
-        .split(chunks[0]); // Aquí puedes usar `chunks[0]` para colocar el bloque en el área adecuada.
-
-    f.render_widget(repo_block, repo_area[0]);
-
-    let description_lines = app.pull_request.description.lines().count();
-    let description_height = description_lines.min(15) + 2;
-    let form_layout = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(2)
-        .constraints(
-            [
-                Constraint::Length(1),
-                Constraint::Length(description_height as u16),
-                Constraint::Length(1),
-                Constraint::Length(1),
-                Constraint::Length(3),
-            ]
-            .as_ref(),
-        )
-        .split(chunks[1]);
-    let form_block = Block::default().title("Create").borders(Borders::ALL);
-    f.render_widget(form_block, chunks[1]);
-    let fields = vec![
-        ("Title", &app.pull_request.title),
-        ("Description", &app.pull_request.description),
-        ("Source Branch", &app.pull_request.source_branch),
-        ("Target Branch", &app.pull_request.target_branch),
-    ];
-
-    for (i, (name, value)) in fields.iter().enumerate() {
-        let (text, style) = match app.input_mode {
-            InputMode::Normal => (
-                format!("{}: {}", name, if value.is_empty() { "" } else { value }),
-                Style::default().fg(if i == app.current_field {
-                    Color::Yellow
-                } else {
-                    Color::White
-                }),
-            ),
-            InputMode::Editing => (
-                format!("{}: {}", name, value),
-                if i == app.current_field {
-                    Style::default().fg(Color::Green)
-                } else {
-                    Style::default().fg(Color::White)
-                },
-            ),
-            InputMode::Creating => (
-                format!("{}: {}", name, value),
-                Style::default().fg(Color::White),
-            ),
-        };
-
-        if i == 1 {
-            let paragraph = Paragraph::new(Text::from(text).style(style))
-                .block(Block::default())
-                .wrap(Wrap { trim: true });
-            f.render_widget(paragraph, form_layout[i]);
-        } else {
-            let paragraph = Paragraph::new(Span::styled(text, style));
-            f.render_widget(paragraph, form_layout[i]);
-        }
-    }
-
-    if let Some(ref error_message) = app.error_message {
-        let error_paragraph = Paragraph::new(Span::from(Span::styled(
-            error_message,
-            Style::default().fg(Color::Red),
-        )))
-        .block(Block::default().borders(Borders::ALL).title("Error"));
-        f.render_widget(error_paragraph, chunks[2]);
-    }
-
-    if let Some(ref success_message) = app.success_message {
-        let success_paragraph = Paragraph::new(Text::from(Text::styled(
-            success_message,
-            Style::default().fg(Color::Green),
-        )))
-        .block(Block::default().borders(Borders::ALL).title("Success"));
-        f.render_widget(success_paragraph, chunks[2]);
-    }
-    // Instructions
-    let instructions = match app.input_mode {
-        InputMode::Normal => "Press 'e' to edit Title, 'c' to create PR, 'q' to quit",
-        InputMode::Editing => "Editing mode. Press 'Esc' to exit, 'Tab' to move to next field",
-        InputMode::Creating => {
-            "Press 'n' to create a new PR, 'e' to edit the PR, 'Enter' to submit"
-        }
-    };
-    let instructions_paragraph =
-        Paragraph::new(instructions).style(Style::default().fg(Color::Gray));
-    f.render_widget(instructions_paragraph, chunks[3]);
-
-    if app.show_popup {
-        let popup_block = Block::default()
-            .title("Pull Request Created")
-            .borders(Borders::ALL);
-
-        let area = centered_rect(60, 20, f.area());
-        f.render_widget(Clear, area);
-        f.render_widget(popup_block, area);
-
-        let popup_text = vec![
-            Line::from(Span::styled(
-                "Pull Request Details:",
-                Style::default().fg(Color::Green),
-            )),
-            Line::from(""),
-            Line::from(format!("Title: {}", app.pull_request.title)),
-            Line::from(format!("Description: {}", app.pull_request.description)),
-            Line::from(format!("Source Branch: {}", app.pull_request.source_branch)),
-            Line::from(format!("Target Branch: {}", app.pull_request.target_branch)),
-            Line::from(""),
-            Line::from(Span::styled(
-                "Press 'n' to create a new PR",
-                Style::default().fg(Color::Yellow),
-            )),
-        ];
-
-        let popup_paragraph = Paragraph::new(popup_text)
-            .block(Block::default().borders(Borders::NONE))
-            .alignment(ratatui::layout::Alignment::Left);
-
-        f.render_widget(popup_paragraph, inner_area(area));
-    }
-}
-
-fn centered_rect(percent_x: u16, percent_y: u16, r: Rect) -> Rect {
-    let popup_layout = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(
-            [
-                Constraint::Percentage((100 - percent_y) / 2),
-                Constraint::Percentage(percent_y),
-                Constraint::Percentage((100 - percent_y) / 2),
-            ]
-            .as_ref(),
-        )
-        .split(r);
-
-    Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints(
-            [
-                Constraint::Percentage((100 - percent_x) / 2),
-                Constraint::Percentage(percent_x),
-                Constraint::Percentage((100 - percent_x) / 2),
-            ]
-            .as_ref(),
-        )
-        .split(popup_layout[1])[1]
-}
-
-fn inner_area(area: Rect) -> Rect {
-    let inner = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(1)
-        .constraints([Constraint::Min(0)].as_ref())
-        .split(area);
-    inner[0]
 }
